feat(BigPlayer): finished edit bookmark

+++ core-network/src/main/java/dev/halim/core/network/request/BookmarkRequest.kt
@@ -2,4 +2,4 @@ package dev.halim.core.network.request
 
 import kotlinx.serialization.Serializable
 
-@Serializable data class BookmarkRequest(val time: Int, val title: String)
+@Serializable data class BookmarkRequest(val time: Long, val title: String)
